c1 coupling, minor improvements
it appears that there is a minimal required value for c1, otherwise the model becomes infeasible. This means that in low growth conditions, the model is infeasible if we couple mrna dilution to protein synthesis rather than mrna formation to protein synthesis. This is likely due to the fact that, as c1 decreases, demand for mrna decreases, and flux through mrna biomass decreases; the model then tries to compensate that by increasing flux through other marcomolecules and becomes infeasible. A few options may improve this: increasing mrna demand by consuming biomass during dilution coupling, by explicitly coupling protein degradation to catalysis, by increasing the number of proteins needed to be synthesize with the introduction of a dummy protein, or by altering the ribosome coupling coefficient